# lxdapi/docs/swagger.yaml
definitions:
...
  lxdapi_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  models.UpdateContainerConfigRequest:
    properties:
      allow_nesting:
        type: boolean
      cpu:
        type: integer
      cpu_allowance:
        type: integer
      disk:
        type: integer
      egress:
        type: integer
      ingress:
        type: integer
      io_read:
        type: integer
      io_write:
        type: integer
      ipv4_mapping_limit:
        type: integer
      ipv4_pool_limit:
        type: integer
      ipv6_mapping_limit:
        type: integer
      ipv6_pool_limit:
        type: integer
      memory:
        type: integer
      memory_swap:
        type: boolean
      privileged:
        type: boolean
      processes_limit:
        type: integer
      remark:
        type: string
      reverse_proxy_limit:
        type: integer
      traffic_limit:
        type: integer
    type: object
info:
  contact: {}
paths:
...
  /api/admin/containers:
    get:
      consumes:
      - application/json
      description: 获取容器列表，可按用户筛选
      parameters:
      - description: 用户ID
        in: query
        name: user_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取容器列表
      tags:
      - Admin API - 容器管理
  /api/admin/containers/:name:
    delete:
      consumes:
      - application/json
      description: 删除指定容器
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除任务已创建
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少参数
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 删除失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 删除容器
      tags:
      - Admin API - 容器管理
    get:
      consumes:
      - application/json
      description: 获取指定容器的详细信息
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少参数
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取容器详情
      tags:
      - Admin API - 容器管理
  /api/admin/containers/:name/action:
    post:
      consumes:
      - application/json
      description: '对容器执行操作: start/stop/restart/pause/resume/reinstall/reset-password/reset-traffic'
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 操作类型
        in: query
        name: action
        required: true
        type: string
      - description: 操作参数
        in: body
        name: request
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 操作失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 容器操作
      tags:
      - Admin API - 容器管理
  /api/admin/containers/:name/config:
    get:
      consumes:
      - application/json
      description: 获取容器的资源配置信息（仅管理员可用）
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少参数
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取容器配置
      tags:
      - Admin API - 容器管理
    put:
      consumes:
      - application/json
      description: 热更新容器配置，支持CPU、内存、磁盘、带宽等升降级，磁盘只能扩容
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 配置更新参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateContainerConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 更新失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 更新容器配置
      tags:
      - Admin API - 容器管理
  /api/admin/containers/:name/credential:
    get:
      consumes:
      - application/json
      description: 获取或创建容器访问凭证
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少参数
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 创建失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取容器凭证
      tags:
      - Admin API - 容器管理
    post:
      consumes:
      - application/json
      description: 重新生成容器访问Hash
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 生成成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少参数
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 生成失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 重新生成容器凭证
      tags:
      - Admin API - 容器管理
  /api/admin/containers/:name/ip:
    get:
      consumes:
      - application/json
      description: 获取容器的IP地址
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 'IP版本: v4/v6/all，默认all'
        in: query
        name: version
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 缺少参数
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 获取容器IP
      tags:
      - Admin API - 容器管理
  /api/admin/containers/:name/ip/allocate:
    post:
      consumes:
      - application/json
      description: 为容器分配IP地址
      parameters:
      - description: 容器名称
        in: path
        name: name
        required: true
        type: string
      - description: 'IP版本: v4 或 v6'
        in: query
        name: version
        required: true
        type: string
      - description: 分配数量
        in: body
        name: request
        schema:
          properties:
            count:
              type: integer
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 分配成功
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "404":
          description: 容器不存在
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
        "500":
          description: 分配失败
          schema:
            $ref: '#/definitions/lxdapi_pkg_response.Response'
      security:
      - SessionAuth: []
      summary: 分配容器IP
      tags:
      - Admin API - 容器管理
...
swagger: "2.0"